feat: move infra unit tests from ./test to ./test/infra
